#  Move emoji parsing