--- a/hugolib/page_content.go
+++ b/hugolib/page_content.go
@@ -17,6 +17,8 @@ import (
 	"bytes"
 	"io"
 
+	"github.com/gohugoio/hugo/helpers"
+
 	errors "github.com/pkg/errors"
 
 	bp "github.com/gohugoio/hugo/bufferpool"
@@ -149,6 +151,12 @@ Loop:
 			result.WriteString(placeHolder)
 			ordinal++
 			s.shortcodes.Add(placeHolder, currShortcode)
+		case it.Type == pageparser.TypeEmoji:
+			if emoji := helpers.Emoji(it.ValStr()); emoji != nil {
+				result.Write(emoji)
+			} else {
+				result.Write(it.Val)
+			}
 		case it.IsEOF():
 			break Loop
 		case it.IsError():
@@ -170,7 +178,10 @@ Loop:
 
 func (p *Page) parse(reader io.Reader) error {
 
-	parseResult, err := pageparser.Parse(reader)
+	parseResult, err := pageparser.Parse(
+		reader,
+		pageparser.Config{EnableEmoji: p.s.Cfg.GetBool("enableEmoji")},
+	)
 	if err != nil {
 		return err
 	}