--- a/parser/pageparser/item.go
+++ b/parser/pageparser/item.go
@@ -113,6 +113,7 @@ const (
 	TypeFrontMatterTOML
 	TypeFrontMatterJSON
 	TypeFrontMatterORG
+	TypeEmoji
 	TypeIgnore // // The BOM Unicode byte order marker and possibly others
 
 	// shortcode items